feat(django,react): Stop exporting default components and move configureAuth (#33)
BREAKING CHANGE: If you used the components exported by default until now, you will
need to create a configuration file, export the components you need, and import
them from that configuration file. See the README for more information.
The configureAuth function is not exported as a toplevel function anymore.
Now you need to import it from 'ai-kit-auth/dist/config/ReactRouter'.
Furthermore, the backend api url cannot passed to the UserStore as a prop anymore,
and is instead defined in the configuration.