feat(django): allow custom registration serializer
This allow setting a custom registration serializer. This includes a small refactoring of the validation method of our registration serializer - the model is now saved in the save method, not in the validation method. This was done initially to get not-unique errors in the validation method, but its confusing and it makes more sense this way.

All tests run with the custom serializer now.